data_collection/requirements.txt
absl-py==2.3.1
attrs==25.3.0
cachetools==5.5.2
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
contourpy==1.3.2
cycler==0.12.1
flatbuffers==25.2.10
fonttools==4.59.0
google-api-core==2.25.1
google-api-python-client==2.176.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
googleapis-common-protos==1.70.0
httplib2==0.22.0
idna==3.10
ImageHash==4.3.2
jax==0.6.2
jaxlib==0.6.2
kiwisolver==1.4.8
matplotlib==3.10.3
mediapipe==0.10.21
ml_dtypes==0.5.1
numpy==1.26.4
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opt_einsum==3.4.0
packaging==25.0
pillow==11.3.0
proto-plus==1.26.1
protobuf==4.25.8
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytube==15.0.0
PyWavelets==1.8.0
PyYAML==6.0.2
requests==2.32.4
rsa==4.9.1
scipy==1.16.0
sentencepiece==0.2.0
six==1.17.0
sounddevice==0.5.2
tqdm==4.67.1
uritemplate==4.2.0
urllib3==2.5.0
